requirements.txt
cadquery>=2.4.0
pyyaml>=6.0
numpy>=1.24.0
